Move /words to a module
